requirements.txt
audiofile==1.0.1
audioread==2.1.9
Keras_Preprocessing==1.1.2
librosa==0.8.1
matplotlib==3.4.3
numpy==1.20.3
pandas==1.3.4
Pillow==8.4.0
pygame==2.1.0
tensorflow==2.4.1
